requirements.txt
PyInquirer
thirdweb-sdk
termcolor
python-dotenv
pyfiglet